feat: v1.0 build complete, exports to xlsx satis
This is the first complete build of the code. It takes two input csv
files, filters them by airport and then runs various investigative
functions on them, before printing the data to the shell and exporting
to an xlsx file in an easy to read format.

It is function based, complies as far as possible with PEP8 and can be
adjusted as required to compare any two csv data files or any of the
reporting airport by easily changing variables.